[Kernels] Remove a redundant wait. (#57675)
I'm surprised this isn't UB. Note that we have a `wait` few lines
earlier (617).
I kept the earlier one because the latter was inside an `unswitch`
(i.e., the latter would generate twice the code).

It'd be nice to delete this other `wait` and the corresponding `arrive`
at the top of the consumer branch.
However, this would require an update to `PipelineState`.

It might actually be simpler, as then `K_{i+1}` and `V_i` would have the
same pipeline state, and these are handled together in a loop iteration.
That would be a more involved code change, so I'm starting simple.

NN_ORIG_REV_ID: 28e9af4bbc1f3f471e9d9a6cab4e5f3ee2879857