[stdlib] Enable debug_assert print with thread info on AMD GPU
This enables printing the message and thread information on AMD GPUs.
Introduced a type named `WritableVariadicPack` that takes a
`VariadicPack` as a reference, only compiling if the types are
`Writable`. This allows forwarding the writable variadics and adding
extra thread and file location information.

This also simplifies the `debug_assert` implementation, allowing reuse
of the buffering functionality from `print`.

Other methods to fix this like using a constraint and rebinding to a
`Writable` variadic pack caused issues, because it loses the concrete
type information of the variadics.

MODULAR_ORIG_COMMIT_REV_ID: 37a19310e53c65322920db5e04beea5985a21e04