[Kernels] Make `nd_buffer_stub` private (#57945)
This module is not intended for the public: it only contains a few
helpers functions used in a few places of non-public APIs internally.
This avoids needing to add API docstrings as we document the `layout`
package. In the future, this module may be removed entirely.

Fixes MSTDL-1280.

QUANTIZATION_ORIG_REV_ID: b155a48098b4a75dbbbe1d10c6085232cf490d6a