[Kernels] Enable API doc validation for `layout` package (#58228)
As the quest of documenting the `layout` package comes to an end, fix
the stray errors and turn on `validate_missing_docs` for the Bazel rule
for the `layout` package. This means that — going forward, public API
docs for things in the `layout` package, must be compliant.

Fixes MSTDL-1307.

LAYOUT_ORIG_REV_ID: 05ffd9330a9a5433adf0837a164bbf4c97cbb088