[MO] Allow mo.while cond block to yield values (#58194)
The cond block of mo.while is currently limited to yielding a single
boolean loop control value. This imposes limitations on what can be done
in the cond block when dealing with operations requiring `!mo.chain`
values.

This change loosens the restrictions on the cond block, requiring it to
yield as many values as the `mo.while` operation has inputs. This brings
the structure of the cond block in line with the body block, and allows
for the use of `!mo.chain` operations in the cond block.

Closes: GEX-1950
MAX_GRAPH_API_ORIG_REV_ID: bb3feeb61cc3124c958b4332443835d55e6836dc