[MOGG][Extensibility] Add extra checks to ParametersMapper to validate kernels. (#57215)
Add new missing parameter checks to ParametersMapper in MoToMOGG:
- Check if multiple uses of a parameter resolve to the same value.
- Check if the resolved value of the parameter is correct for constants (eg InputTensor[rank=4]).

We want to run these checks during the Graph building API, but it's easier as a first step to run them on MoToMOGG pass.
In a future PR this part of the code will be refactored slightly to be used by both MoToMOGG and the Graph API (In the Graph API case we won't actually build the parameters and repametrize the kernels, we'll only infer and check the parameter values).

Towards GEX-1800.

MOGG_ORIG_REV_ID: db8fcc1b8216505cf3afb5b68b4e9fa1f58e120f