[Kernels][AMD] update demo matmul kernel (#58333)
Changes:
1. Update the shared memory layout to not rely on nested tensors when
adapted to layout tensors. The swizzling pattern relied on the outer
BNxBK shape as of the last change, but I rearranged the layout again to
make the swizzling only based on the local MNxMK tile. This should
hopefully help in the translation back to layout tensors and there was
no loss in performance from the shared memory changes.
2. The other motivation for the above was to implement experimental
support for the 32x32x8 MFMA instruction. The kernel has been updated to
fully support this format as well with no bank conflicts. The reason I
cared is that rocblas/hipblaslt/aiter dispatch to kernels using this
shape especially for medium sized M. Early data though isn't super
compelling, so need to adapt this for the layout tensor version yet: I
haven't found a shape yet where it beats the existing 16x16x16 kernel.
3. Removed the experiment of using DWORD loads from global memory. This
was implemented to try to mirror Tensile generated kernels, but there
was no payoff and maintaining that code was painful.
GPU_ORIG_REV_ID: a77ea7dddafd2794370ed87e3bd84b26f400499a