[Kernels][MOGG] Hook Up Runtime Shape Function For Advanced Indexing (#58302)
This was forgotten to be done. Luckily it is easy as a utility was made for testing in pure Mojo.

This turns out to be needed in some PyTorch cases. This is also ripe for some general algebraic refinement which we don't have a design for yet.

MOGG_ORIG_REV_ID: fb7ce31ea83637bd8fc9f1b7ca8ce5a4fc656361