[Kernels] Add argsort kernel (#58950)
This adds a CPU and GPU implementation of the argsort kernel. It's not
meant to be super optimized, but it's a good starting point. For the
CPU, the argsort will use our existing sort function. For the GPU
implementation, the argsort uses the Bitonic Sort algorithm. It sorts an
array on the GPU and returns the indices that would sort the array in
ascending order. Currently this is limited to 1D tensors.

MOGG_ORIG_REV_ID: 8bbf1355bd74a10f4b3e319138496f924c66dcea