[mojo-lang] Decouple `def` and `object`
This changes processing of `def` to default to returning `None`
instead of `object` when a return type is not specified, and
requiring argument types on def arguments (instead of defaulting
to `object`).

We can look into restoring this behavior in the future as Mojo
shifts to wanting to support more dynamic and fluid programming
styles, but right now the `object` type is underdeveloped and
should be removed.

With this change, the Mojo compiler doesn't know about `object`
at all.

This fixes MOCO-1778

MODULAR_ORIG_COMMIT_REV_ID: f7b2ea4c192845a0530b34ce4c9ac70c6e838426