[mojo-stdlib] Remove the join/split methods from StringLiteral.
This PR sinks the join method on StringLiteral to StringSlice,
which is a more general place to support this.  It removes join
and split from StringLiteral, because we want StringLiteral to
eventually autopromote to string or StaticString instead of
redundantly implementing all of these.

MODULAR_ORIG_COMMIT_REV_ID: c0d6466e3ba2ffd2c2d583ce84a88986f75361cb