Revert "[Kernels][AMD] Use 256x256x64 block sizes for M,N, K >= 4096" (#59229)
Reverts [Internal link]

Performance is down as seen here:
https://modular.cloud.looker.com/looks/286. One possible fix is #59228
and the other is just reverting the change until the degradation is
understood. Benchmarks are going to be run on this revert to see if that
improves the benchmarks.

LINALG_ORIG_REV_ID: f0afbae92597e8a3ab312307256ccdb09f4f659f