[serve] Remove max/serve/grpc_server (#59359)
The GRPC Server is an old request that has not been touched in a long
time. As such, it is stale and not specifically needed anymore. On top
of that, it makes many huggingface requests. We are trying to remove
huggingface requests to reduce 429 errors. Decided to just remove the
GRPC server.

Fixes AITLIB-309

SERVE_ORIG_REV_ID: b1b73642109211fba4f7d952ead2ec13d33af281