[mojo-lang][CheckLifetime] Fix checkConsume with trivial types. (#59626)
This fixes a bug in handling of consume markers (which are generated in
EH regions) that caused us to fail to reject the example.  This turned
up as a bogus warning on test_arange.mojo with the dead store checker.

NN_ORIG_REV_ID: 9d763650b0d04ad95831ca99ee103b3250a5d097