[SDK] Add optional device to ops.range (#59653)
Add an optional device argument to `ops.range`.
This op is interesting in that all of its operands are scalars, which
should be on CPU.
So it's not possible to infer the result device from the operands.
For that reason, give it a device, which is optional for now to
enable incrementally opting in to explicit device.

MAX_GRAPH_API_ORIG_REV_ID: d14742da6ad7e28678626fe3433efb68c95bff7b