[Kernels][gpu] use larger smem tile for TMA store (#59225)
Currently we clip WG_BN to 64 because of swizzling constrained for TMA
store. This PR extend this to the original 128. We still clip each TMA
store to BNx64 for each TMA store, but the c smem is larger so we save
some time for STSM synchronization.

Part of KERN-1713

GPU_ORIG_REV_ID: e697bb07dfa100bef81d348209e44c89722884f7