[Stdlib] Add a vectorized ArcTanh implementation
This implements a vectorized ArcTanh implementation that uses a
polynomial approximation for the range [0, 0.5] and a logarithmic
identity for the range [0.5, 1). This is more performant and portable
than the existing implementation which uses the libm function.

MODULAR_ORIG_COMMIT_REV_ID: 7358fc3f2cba15ff1247c0e6e0fbb75320eaad69