[mojo-stdlib] Implement constant string optimization
This changes string to not allocate memory when pointing to a
known-static-constant string like `StringLiteral` and `StaticString`.
We can simply point to the existing data and not free it when we are
destroyed.

In addition to improving efficiency, this allows us to make
conversions from these types be implicit constructions which is far
more ergonomic for common cases.

As part of doing this, we need to be more careful about mutable
accesses to strings - getting a mutable pointer into a string may
need to copy the data to ensure it is safe to change.  As such, this
splits various ref methods like `unsafe_ptr()` to have an explicit
`unsafe_ptr_mut()` method since most accesses into the string
do not need the copy or mutabilty.

NN_ORIG_REV_ID: b0005c3a232d5f31042dc9251865cb84a317a536