[SDK] Use rmo.conv_transpose in the graph API for ops.conv2d_transpose (#60184)
Convert ops.conv2d_tranpose to use rmo.conv_transpose. The benefit of
this is that the code which infers the result type of the
ops.conv2d_transpose operation is shared by the logic in
shape-propagation. This structure is similar to how conv2d and conv3d
are implemented in the graph API.

Closes: GEX-2129
MAX_GRAPH_API_ORIG_REV_ID: b55693457bfb21b250221dd4c248dc4df2e3b007