[SDK] fix: Remove `;` in rpath of binaries built with max.entrypoints.mojo runner #60366
The max.entrypoints.mojo package is a way to invoke the `mojo`
executable within an environment configured to use the various data
files that are present in package SDK when installed on end-user
machines.

This extra ";" was causing the linker to be given the flags:

    -Xlinker rpath -XLinker /some/lib/dir;

as arguments, which the linker then ignored because the specified
directory with a semicolon on the end did not exist.

The max.entrypoints.mojo code is not currently widely used, so this was
caught while trying to add a new _core_mojo module to the SDK sources in
#59909.

Co-authored-by: Alex Trotta <44127594+Ahajha@users.noreply.github.com>
MODULAR_ORIG_COMMIT_REV_ID: 36a8aca9c00d475b0df1fdb6c97f55e22149056f