[RecordReplay] Add read timeout option to replay_recording
The default value of two minutes can be too low in some scenarios.
Typically this means the concurrency of requests is higher than the
server can handle in its current batch configuration, but it's best to
be able to adjust this to test suboptimal scenarios as well.

MODULAR_ORIG_COMMIT_REV_ID: 638fa55906d921f2d3692bebee16d65f2998ec72