[stdlib] Add docs for `runtime.tracing`
The `runtime` package will be moving into the `stdlib` and be checked
for strict API docs shortly. Help pave the way for this by adding API
docs for `runtime.tracing`.

This module needs a fresh look as to what we should be exposing to the
public, but I'll leave that for someone else in a follow up.

RUNTIME_ORIG_REV_ID: dc511d7896a2bc53ede6fd34e801273fcf6afc5f