[SDK] Insert constant.external ops at the top of the graph body block (#59912)
This change adds the ability to specify an insertion point when adding an op to the graph. This is then used to insert `mo.constant.external` ops emitted by `Weight` instances at the top of the graph. This allows for the same `mlir.Value` instances to be hoisted out of control flow ops, which will allow us in an upcoming PR to cache the `mlir.Value` instance, even when evaluated in sub-regions.

MAX_GRAPH_API_ORIG_REV_ID: 4da2a1be642f436d9f1f713bcbf0053146aeee65