[Kernels] add static strides to KVCache buffers (#60493)
Add static shape and more importantly static stride information to the
buffers in the KVCache object. This reduces the code size of the QKV
matmul kernel on AMD where currently every element is written as a
scalar and all of the KVCache pointer math was generating a huge number
of conditional selects. Roughly a 10% reduction in that kernel size.
This should help a few other kernels and other architectures benefit
from this.

KV_CACHE_ORIG_REV_ID: 2cb140ac40fbbedef56832afea2cec279810fb7a