[mojo-lang] Tighten param call emission.
The parser has logic to determine when (in a dynamic context) to
force-emit
an expression as a parameter expression.  For example, when we have
something like: `var x = True`, we don't need to emit a dynamic call to
`Bool.init(i1 1)`, we can evaluate this at comptime and just get the
result.

The previous implementation of this did this transformation when all
arguments are comptime, and when (a: callee is `always_inline(builtin)`)
or (b: when any argument is a nonmaterializable). The "A" case is good,
but
the "b" case is bad because it can omit side effects.

This removes the b case, tightening things up. This fixes
https://github.com/modular/max/issues/4362
and fixes the issue behind #59819 that was causing us to comptime
evaluate
random python stuff (and fail).

MODULAR_ORIG_COMMIT_REV_ID: b9527c00fa00ebe778beac659a9754c4265c9a09