[SDK][Graph API] Make Weight Sharding Explicit (#60601)
Stacked PRs:
 * #59300
 * #60600
 * #60599
 * __->__#60601

--- --- ---

### [SDK][Graph API] Make Weight Sharding Explicit

There is no practical usage of using "None" as the weight shard (which
will use the native device in the original weight object), so remove it.

This is also more in-line with the rest of the API which is explicit.

MAX_GRAPH_API_ORIG_REV_ID: 5b4e5cfcebdbbe64f51951788ecd8cf0e5209840