[SDK] Improve op addition to graph when not appending to block end (#60645)
This change improves the behavior of the `Graph` class when adding an
operation to the graph that is not at the end of the block.

The previous attempt at this did not compose well when the `op` function
returned instances of `max.graph.Value` instead of `mlir.Value`
instances.

MAX_GRAPH_API_ORIG_REV_ID: 417529ed5d4b10167c40f68475671d9166665f63