[stdlib] Load cuBLAS libraries by name When loading cuBLAS and cuBLASLt libraries, try the library name first, before any absolute library paths. This way, if PyTorch already loaded the library, we will use that instead of loading another possibly conflicting version. Additionally, if the library is installed in a non-standard path, the system linker ought to be able to find it in the library cache.
Also require CUDA version 12, we shouldn't allow loading an older
version.

In case of an absolute path that does not exist, we now invoke dlopen()
instead of skipping it. But this should have the same effect as before,
we just catch the error and move on.

MODULAR_ORIG_COMMIT_REV_ID: 19bcec22de40e1b43bac1cb1095711d1b0159fce