[SDK] Validate number of inputs in call to subgraph (#60924)
Added a check in the `ops.call` function to ensure the number of provided arguments matches the expected input count for the subgraph. This prevents runtime errors due to mismatched input sizes.

Introduced a new test case to verify that a ValueError is raised when there is a mismatch in the number of inputs.

Towards GEX-2017

MAX_GRAPH_API_ORIG_REV_ID: bc9c029f6fe1f0eb183cceb65071bd9a0556a5db