[Stdlib] Use `vectorize` in functions instead of reimplementing SIMD code
The use of `vectorize` is more idiomatic and easier to understand than
reimplementing SIMD code. We were unable to do this before because the
algorithm package was not oss. There are more patterns like this that
can be applied, but this is a good start.

MODULAR_ORIG_COMMIT_REV_ID: 5d295f45148406d25b5b1e302f3f374f7924e355